refactor how code cells save
