+++ quadratic-api/src/middleware/getFile.ts
@@ -39,8 +39,10 @@ export async function getFile<T extends number | undefined>({ uuid, userId }: { 
   }
 
   const isFileOwner = !file.ownerTeamId && file.ownerUserId === userId;
-  const teamRole = file.ownerTeam && file.ownerTeam.UserTeamRole[0] ? file.ownerTeam.UserTeamRole[0].role : undefined;
-  const fileRole = file.UserFileRole[0] ? file.UserFileRole[0].role : undefined;
+
+  // todo...
+  const teamRole = undefined; //file.ownerTeam && file.ownerTeam.UserTeamRole[0] ? file.ownerTeam.UserTeamRole[0].role : undefined;
+  const fileRole = undefined; //file.UserFileRole[0] ? file.UserFileRole[0].role : undefined;
 
   const filePermissions = getFilePermissions({
     fileRole,
